Fix comments from review Change Package Name Use LiveData and viewModel Refactor movie trailer view
