+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.moviesapp">
+    package="com.ehnyn.moviesapp">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -12,9 +12,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.MovieReviewActivity"></activity>
-        <activity android:name=".activities.DetailsActivity" />
-        <activity android:name=".activities.MainActivity">
+        <activity android:name="com.ehnyn.moviesapp.activities.MovieReviewActivity"></activity>
+        <activity android:name="com.ehnyn.moviesapp.activities.DetailsActivity" />
+        <activity android:name="com.ehnyn.moviesapp.activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
