+++ app/src/main/java/com/ehnyn/moviesapp/models/Database/MovieDao.java
@@ -1,0 +1,36 @@
+package com.ehnyn.moviesapp.models.Database;
+
+import android.arch.lifecycle.LiveData;
+import android.arch.persistence.room.Dao;
+import android.arch.persistence.room.Delete;
+import android.arch.persistence.room.Insert;
+import android.arch.persistence.room.OnConflictStrategy;
+import android.arch.persistence.room.Query;
+import android.arch.persistence.room.Update;
+
+import com.ehnyn.moviesapp.models.MoviesResult;
+
+import java.util.List;
+
+@Dao
+public interface MovieDao {
+
+    @Query("Select * from movie order by movieId")
+    List<MoviesResult> fetchAllMoviesById();
+
+
+    @Query("Select * from movie")
+    LiveData<List<MoviesResult>> fetchAllMovies();
+
+    @Query("Select * from movie where 'starred' = 'true'")
+    List<MoviesResult> getAllStarredMovies();
+
+    @Insert
+    void saveMovie(MoviesResult moviesResult);
+
+    @Update(onConflict = OnConflictStrategy.REPLACE)
+    void updateMovie(MoviesResult moviesResult);
+
+    @Delete
+    void deleteMovie(MoviesResult moviesResult);
+}
