+++ app/src/main/java/com/ehnyn/moviesapp/models/OnItemClickedListener.java
@@ -1,0 +1,6 @@
+package com.ehnyn.moviesapp.models;
+
+public interface OnItemClickedListener {
+
+    void onItemClicked(MoviesResult movieResult);
+}
