--- app/src/main/java/com/example/moviesapp/models/MoviesModel.java
@@ -1,57 +1,0 @@
-package com.example.moviesapp.models;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class MoviesModel {
-
-    @SerializedName("page")
-    @Expose
-    private int page;
-
-    @SerializedName("total_results")
-    @Expose
-    private int totalResults;
-
-    @SerializedName("total_pages")
-    @Expose
-    private int totalPages;
-
-    @SerializedName("results")
-    @Expose
-    private List<MoviesResult> results;
-
-    public int getPage() {
-        return page;
-    }
-
-    public void setPage(int page) {
-        this.page = page;
-    }
-
-    public int getTotalResults() {
-        return totalResults;
-    }
-
-    public void setTotalResults(int totalResults) {
-        this.totalResults = totalResults;
-    }
-
-    public int getTotalPages() {
-        return totalPages;
-    }
-
-    public void setTotalPages(int totalPages) {
-        this.totalPages = totalPages;
-    }
-
-    public List<MoviesResult> getResults() {
-        return results;
-    }
-
-    public void setResults(List<MoviesResult> results) {
-        this.results = results;
-    }
-}
