--- app/src/main/java/com/example/moviesapp/models/OnItemClickedListener.java
@@ -1,6 +1,0 @@
-package com.example.moviesapp.models;
-
-public interface OnItemClickedListener {
-
-    void onItemClicked(MoviesResult movieResult);
-}
