--- app/src/main/java/com/example/moviesapp/networking/RetrofitClient.java
@@ -1,49 +1,0 @@
-package com.example.moviesapp.networking;
-
-import com.example.moviesapp.BuildConfig;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
-import okhttp3.OkHttpClient;
-import okhttp3.logging.HttpLoggingInterceptor;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-public class RetrofitClient {
-
-    private static Retrofit retrofit = null;
-
-    public Retrofit getRetrofit(String baseUrl) {
-
-        if(retrofit == null){
-            Gson gson = new GsonBuilder()
-                    .setLenient()
-                    .create();
-
-            retrofit = new Retrofit.Builder()
-                    .baseUrl(baseUrl)
-                    .client(okHttpClient())
-                    .addConverterFactory(GsonConverterFactory.create(gson))
-                    .build();
-        }
-
-        return retrofit;
-    }
-
-    private OkHttpClient okHttpClient(){
-        OkHttpClient.Builder okHttpClient = new OkHttpClient.Builder();
-
-        HttpLoggingInterceptor loggingInterceptor = new HttpLoggingInterceptor();
-
-        if(BuildConfig.DEBUG){
-            loggingInterceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
-        } else {
-            loggingInterceptor.setLevel(HttpLoggingInterceptor.Level.BASIC);
-        }
-
-        okHttpClient.addInterceptor(loggingInterceptor);
-
-        return okHttpClient.build();
-    }
-
-}
