+++ app/src/main/res/layout/activity_details.xml
@@ -32,18 +32,20 @@
     <android.support.v7.widget.CardView
         android:id="@+id/movieDetailCardView"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="350dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/movieThumbnail">
 
         <android.support.constraint.ConstraintLayout
             android:id="@+id/movieInformationCardView"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="match_parent"
+            android:layout_marginBottom="16dp"
+            android:layout_marginTop="16dp">
 
             <ImageView
                 android:id="@+id/movie_poster"
-                android:layout_width="200dp"
+                android:layout_width="220dp"
                 android:layout_height="300dp"
                 tools:ignore="ContentDescription"
                 android:layout_marginStart="12dp"
@@ -56,22 +58,21 @@
                 android:id="@+id/originalTitle"
                 android:layout_height="50dp"
                 android:layout_width="wrap_content"
-                android:layout_marginStart="16dp"
-                android:textSize="16sp"
-                android:padding="8dp"
+                android:layout_margin="8dp"
+                android:textSize="20sp"
                 android:typeface="serif"
+                android:textColor="@color/colorPrimary"
                 app:layout_constraintTop_toTopOf="@+id/movie_poster"
                 app:layout_constraintStart_toEndOf="@id/movie_poster" />
-
 
             <TextView
                 android:id="@+id/releaseDate"
                 android:layout_height="50dp"
                 android:layout_width="wrap_content"
-                android:layout_marginStart="16dp"
-                android:padding="8dp"
-                android:textSize="16sp"
+                android:layout_margin="8dp"
+                android:textSize="20sp"
                 android:typeface="serif"
+                android:textColor="@color/colorPrimary"
                 app:layout_constraintTop_toBottomOf="@+id/originalTitle"
                 app:layout_constraintStart_toEndOf="@id/movie_poster" />
 
@@ -79,21 +80,19 @@
                 android:id="@+id/rating"
                 android:layout_height="50dp"
                 android:layout_width="wrap_content"
-                android:layout_marginStart="16dp"
-                android:textSize="16sp"
-                android:padding="8dp"
+                android:layout_margin="8dp"
+                android:textSize="20sp"
+                android:textColor="@color/colorPrimary"
                 app:layout_constraintTop_toBottomOf="@+id/releaseDate"
                 android:typeface="serif"
                 app:layout_constraintStart_toEndOf="@id/movie_poster" />
-
 
             <ImageView
                 android:id="@+id/starMovie"
                 android:layout_height="50dp"
                 android:layout_width="50dp"
-                android:layout_marginStart="16dp"
-                android:src="@drawable/star_movie_icon"
-                android:background="#ebc834"
+                android:layout_margin="8dp"
+                android:src="@drawable/star_icon"
                 app:layout_constraintTop_toBottomOf="@+id/rating"
                 app:layout_constraintStart_toEndOf="@id/movie_poster"
                 tools:ignore="ContentDescription"/>
@@ -101,32 +100,33 @@
             <Button
                 android:id="@+id/reviewContent"
                 android:layout_height="50dp"
-                android:layout_width="wrap_content"
+                android:layout_width="150dp"
                 android:text="@string/view_review"
                 android:gravity="center"
                 android:typeface="serif"
-                android:background="#26a69a"
-                android:layout_marginStart="16dp"
+                android:textColor="@android:color/white"
+                android:background="@drawable/round_button"
+                android:layout_margin="8dp"
                 android:layout_marginTop="16dp"
                 app:layout_constraintTop_toBottomOf="@id/starMovie"
                 app:layout_constraintStart_toEndOf="@id/movie_poster"/>
-
 
         </android.support.constraint.ConstraintLayout>
 
     </android.support.v7.widget.CardView>
 
         <TextView
-        android:id="@+id/plotSynopsis"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:layout_margin="16dp"
-        android:textSize="16sp"
-        android:padding="8dp"
-        app:layout_constraintTop_toBottomOf="@+id/movieDetailCardView"
-        app:layout_constraintStart_toStartOf="parent"
-        android:layout_marginTop="16dp"
-        android:layout_marginBottom="16dp"/>
+            android:id="@+id/plotSynopsis"
+            android:layout_height="wrap_content"
+            android:layout_width="match_parent"
+            android:layout_margin="16dp"
+            android:textSize="20sp"
+            android:padding="8dp"
+            android:textColor="@color/colorPrimary"
+            app:layout_constraintTop_toBottomOf="@+id/movieDetailCardView"
+            app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginTop="16dp"
+            android:layout_marginBottom="16dp"/>
 
 
         <TextView
@@ -134,9 +134,10 @@
             android:layout_height="wrap_content"
             android:layout_width="match_parent"
             android:layout_margin="16dp"
-            android:textSize="16sp"
+            android:textSize="20sp"
             android:padding="8dp"
             android:text="@string/movie_trailer"
+            android:textColor="@color/colorPrimary"
             android:typeface="serif"
             app:layout_constraintTop_toBottomOf="@+id/plotSynopsis"
             app:layout_constraintStart_toStartOf="parent"
