+++ app/src/main/res/layout/movie_trailer.xml
@@ -1,33 +1,39 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<ScrollView
     android:layout_width="wrap_content"
-    android:layout_height="120dp"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    android:layout_height="wrap_content"
+    android:orientation="horizontal"
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <ImageView
-        android:id="@+id/playButton"
-        android:layout_width="150dp"
-        android:layout_height="100dp"
-        android:layout_margin="0dp"
-        android:adjustViewBounds="true"
-        android:src="@drawable/icons_play_button"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:ignore="ContentDescription"/>
-
-    <TextView
-        android:id="@+id/movieTrailerName"
-        android:layout_height="100dp"
+    <android.support.constraint.ConstraintLayout
         android:layout_width="wrap_content"
-        android:textSize="16sp"
-        android:padding="8dp"
-        android:gravity="center"
-        android:text="@string/movie_trailer"
-        android:typeface="serif"
-        app:layout_constraintStart_toEndOf="@+id/playButton"
-        app:layout_constraintTop_toTopOf="parent"/>
+        android:layout_height="wrap_content"
+        xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        tools:ignore="ContentDescription">
+
+        <ImageView
+            android:id="@+id/trailerThumbnail"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="8dp"
+            android:adjustViewBounds="true"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
 
 
-</android.support.constraint.ConstraintLayout>
+        <ImageView
+            android:id="@+id/playButton"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:src="@drawable/icons_play_button"
+            android:gravity="center"
+            android:padding="8dp"
+            app:layout_constraintStart_toStartOf="@id/trailerThumbnail"
+            app:layout_constraintEnd_toEndOf="@id/trailerThumbnail"
+            app:layout_constraintTop_toTopOf="@id/trailerThumbnail"
+            app:layout_constraintBottom_toBottomOf="@id/trailerThumbnail" />
+
+    </android.support.constraint.ConstraintLayout>
+
+</ScrollView>
