+++ build.gradle
@@ -8,6 +8,7 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.4.2'
+        classpath 'com.neenbedankt.gradle.plugins:android-apt:1.8'
         
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -19,7 +20,6 @@ allprojects {
         google()
         jcenter()
         mavenCentral()
-        
     }
 }
